backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
requests==2.32.3
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0

